requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
pillow
sqlalchemy
reportlab
python-multipart
openai
